refactor Amber simulation wrapper
