--- docs/amber.rst
@@ -1,7 +1,0 @@
-AMBER Simulation Wrapper API
-----------------------------
-
-.. automodule:: paprika.amber
-   :members:
-   :undoc-members:
-
