+++ docs/simulation.rst
@@ -1,0 +1,9 @@
+Simulation Wrappers API
+----------------------------
+
+.. automodule:: paprika.simulate
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
+
